Lista10-04

+++ src/main/java/br/unicesumar/ads/Exercicio2d.java
@@ -9,13 +9,13 @@ public class Exercicio2d implements InterfaceExercicio {
         int num1, num2, num3;
 
         System.out.print("Digite 1° valor: ");
-	num1 = valor.nextDouble();
+	num1 = valor.nextInt();
 		
 	System.out.print("Digite 2° valor: ");
-	num2 = valor.nextDouble();
+	num2 = valor.nextInt();
 		
 	System.out.print("Digite 3° valor: ");
-	num3 = valor.nextDouble();
+	num3 = valor.nextInt();
 
         String strnum1=String.format("%d",num1);
         String strnum2=String.format("%d",num2);
